"remodeling whole functionality"

+++ employeeservImplementation/src/main/java/com/paypal/bfs/test/employeeserv/repositories/IRepository.java
@@ -6,5 +6,5 @@ import java.sql.SQLException;
 
 public interface IRepository {
     Boolean persist(Employee emp) throws SQLException, NullPointerException;
-    Employee retrive(Long id) throws SQLException, NullPointerException;
+    Employee retrive(Integer id) throws SQLException, NullPointerException;
 }
